Backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
cloudinary==1.41.0
dj-database-url==2.3.0
Django==5.1.4
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==1.7.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
